Validate experiment phase prerequisites; share the parent check as a service
The prerequisite-parent check moves out of the features API directory into
services/prerequisiteParents so experiments can use it. Experiment phase
prerequisites now get the same existence, archived, and boolean checks on
the REST create/update endpoints and the app API create, phase edit, new
phase, and targeting writes. Experiments are leaves of the prerequisite
graph, so no cycle walk runs for them. As with features, only parents a
write adds are checked.
