+++ packages/back-end/test/api/features/validations.test.ts
@@ -1,20 +1,13 @@
-import type { FeatureInterface } from "shared/types/feature";
 import {
-  assertValidPrerequisiteParents,
   assertValidRuleEnvironments,
   validateRuleAttributes,
   validateRulesReferences,
 } from "back-end/src/api/features/validations";
 import { BadRequestError } from "back-end/src/util/errors";
 import { ApiReqContext } from "back-end/types/api";
-import { getAllFeaturesWithoutEditorFields } from "back-end/src/models/FeatureModel";
 
 jest.mock("back-end/src/models/FeatureModel", () => ({
   getAllFeaturesWithoutEditorFields: jest.fn(),
-}));
-const scanContext = { scan: true } as unknown as ApiReqContext;
-jest.mock("back-end/src/services/organizations", () => ({
-  getContextForAgendaJobByOrgObject: () => scanContext,
 }));
 
 // `validateRuleAttributes` is the V2-side gate for the opt-in
@@ -250,55 +243,5 @@ describe("assertValidRuleEnvironments", () => {
         { allEnvironments: false, environments: ["prodution"] },
       ]),
     ).toThrow(BadRequestError);
-  });
-});
-
-describe("assertValidPrerequisiteParents", () => {
-  const ctx = { org: { id: "org" } } as unknown as ApiReqContext;
-  const flag = (id: string, ...parents: string[]): FeatureInterface =>
-    ({
-      id,
-      valueType: "boolean",
-      archived: false,
-      environmentSettings: { production: { enabled: true } },
-      rules: [],
-      prerequisites: parents.map((p) => ({ id: p, condition: "{}" })),
-    }) as unknown as FeatureInterface;
-  // Direct parents load with the caller's context; ancestors with the scan
-  // context. Resolves requested ids against whichever map the context gets.
-  const stub = (
-    visible: (context: unknown) => Record<string, FeatureInterface>,
-  ) =>
-    jest
-      .mocked(getAllFeaturesWithoutEditorFields)
-      .mockImplementation(async (context, opts) => {
-        const byId = visible(context);
-        return (opts?.ids ?? []).flatMap((id) => (byId[id] ? [byId[id]] : []));
-      });
-
-  beforeEach(() => jest.mocked(getAllFeaturesWithoutEditorFields).mockReset());
-
-  it("does not query when the write adds no prerequisite", async () => {
-    await assertValidPrerequisiteParents(ctx, flag("c", "p"), flag("c", "p"));
-    expect(getAllFeaturesWithoutEditorFields).not.toHaveBeenCalled();
-  });
-
-  it("finds a cycle through an ancestor the caller cannot read", async () => {
-    // c -> p -> hidden -> c; `hidden` only resolves for the scan context.
-    const p = flag("p", "hidden");
-    const hidden = flag("hidden", "c");
-    stub((context) => (context === scanContext ? { p, hidden } : { p }));
-    await expect(
-      assertValidPrerequisiteParents(ctx, flag("c", "p")),
-    ).rejects.toThrow(/circular dependency/);
-  });
-
-  it("refuses a chain still open after the depth limit", async () => {
-    const byId: Record<string, FeatureInterface> = {};
-    for (let i = 0; i < 60; i++) byId[`n${i}`] = flag(`n${i}`, `n${i + 1}`);
-    stub(() => byId);
-    await expect(
-      assertValidPrerequisiteParents(ctx, flag("c", "n0")),
-    ).rejects.toThrow(/too deep/);
   });
 });
